Move `actions::Comparison` to dedicated submodule
