+++ src/action/mod.rs
@@ -2,6 +2,7 @@
 mod action;
 pub mod actions;
 mod command;
+mod comparison;
 mod handler;
 mod io;
 mod publisher;
@@ -9,6 +10,7 @@ mod routine;
 
 pub use action::{Action, BoxedAction};
 pub use command::*;
+pub use comparison::Comparison;
 pub use handler::SchedRoutineHandler;
 pub use io::IOCommand;
 pub use publisher::Publisher;
